Modified the ghazale_sds_front_end_search_field function
This function previously returned a separate table for each record that
matched the search query in the database.  I made it so that it returned
a single table with a row for each matching record.  Also, I added an
additional search field and the necessary logic for the user to search
by name and by ID (both required).